+++ CHANGELOG.md
@@ -33,6 +33,7 @@ From 0.5.0 onwards, this should be better documented...
 ### Rtvals
 - The response now starts with the current datetime
 - No longer show ungrouped when there's a grouped textval
+- IntegerVals now show up in the rtvals listing
 
 ### Fixes
 - Generics in a path got the id from the filename instead of the id